requirements.txt
# Deep Learning
torch==2.0.1
torchvision==0.15.2
grad-cam==1.4.8

# Processamento de Dados
numpy==1.24.3
pandas==2.0.2
pillow==10.0.0
opencv-python==4.8.0.76
albumentations==1.3.1

# Avaliação e Análise
scikit-learn==1.3.0
scipy==1.11.2

# Visualização
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.0

# Experimentação e Utilitários
tensorboard==2.13.0
pyyaml==6.0.1
tqdm==4.66.1
thop>=0.1.1.post2207130030

# Jupyter (opcional, para notebooks)
jupyter==1.0.0
ipywidgets==8.1.0